Split preprocessor expression code
